Rename the datastore-utils module to core-utils.
