+++ datastore/core-utils/src/main/groovy/org/orbisgis/datastore/coreutils/DataStoreFinderUtils.groovy
@@ -34,38 +34,21 @@
  * or contact directly:
  * info_at_ orbisgis.org
  */
-package org.orbisgis.datastore.datastoreutils
+package org.orbisgis.datastore.coreutils
 
-import groovy.transform.stc.ClosureParams
-import groovy.transform.stc.FirstParam
-import org.geotools.feature.FeatureCollection
-import org.geotools.feature.FeatureIterator
-import org.opengis.feature.Feature
-import org.opengis.feature.type.FeatureType
+import org.geotools.data.DataStore
+import org.geotools.data.DataStoreFinder
+
 
 /**
- * Utility script used as extension module adding methods to {@link org.geotools.feature.FeatureCollection} class.
+ * Utility script used as extension module adding methods to {@link org.geotools.data.DataStoreFinder} class.
  *
  * @author Erwan Bocher (CNRS 2020)
  * @author Sylvain PALOMINOS (UBS chaire GEOTERA 2020)
  */
 
-/**
- * Iterates through an FeatureIterator, passing each item to the given closure.
- *
- * @param fc      The {@link org.geotools.feature.FeatureCollection} over which we iterate.
- * @param closure The closure applied on each element found.
- */
-static <T extends FeatureType, F extends Feature> void each(FeatureCollection<T, F> fc,
-                                                @ClosureParams(FirstParam.FirstGenericType.class) Closure closure) {
-    fc.features().each(closure)
-}
-
-/**
- * Give to {@link org.geotools.feature.FeatureCollection#features()} a more readable name.
- * @param fc FeatureCollection to iterate.
- * @return A FeatureIterator.
- */
-static <T extends FeatureType, F extends Feature> FeatureIterator<F> getFeatureIterator(FeatureCollection<T, F> fc) {
-    fc.features()
+static DataStore getDataStore(DataStoreFinder finder, LinkedHashMap map) {
+    def tmp = [:]
+    map.each {tmp.put(it.key, it.value in GString ? it.value.toString() : it.value)}
+    return DataStoreFinder.getDataStore((Map)tmp)
 }
