+++ datastore/core-utils/src/main/resources/META-INF/groovy/org.codehaus.groovy.runtime.ExtensionModule
@@ -1,0 +1,7 @@
+moduleName=OrbisGIS Groovy core extension
+moduleVersion=1.0
+extensionClasses=org.orbisgis.datastore.coreutils.DataStoreUtils, \
+  org.orbisgis.datastore.coreutils.FeatureIteratorUtils, \
+  org.orbisgis.datastore.coreutils.FeatureCollectionUtils, \
+  org.orbisgis.datastore.coreutils.FeatureSourceUtils
+staticExtensionClasses=org.orbisgis.datastore.coreutils.DataStoreFinderUtils
