+++ datastore/core-utils/src/test/groovy/org/orbisgis/datastore/coreutils/FeatureCollectionUtilsTest.groovy
@@ -1,6 +1,6 @@
-package org.orbisgis.datastore.datastoreutils
+package org.orbisgis.datastore.coreutils
 
-import org.geotools.data.FeatureSource
+
 import org.geotools.data.shapefile.ShapefileDataStore
 import org.junit.jupiter.api.Test
 /**
