+++ datastore/core-utils/src/test/groovy/org/orbisgis/datastore/coreutils/FeatureSourceUtilsTest.groovy
@@ -1,4 +1,4 @@
-package org.orbisgis.datastore.datastoreutils
+package org.orbisgis.datastore.coreutils
 
 import org.geotools.data.Query
 import org.geotools.data.shapefile.ShapefileDataStore
