+++ datastore/core-utils/src/test/resources/org/orbisgis/datastore/coreutils/landcover2000.shp
