--- datastore/datastore-utils/src/main/resources/META-INF/groovy/org.codehaus.groovy.runtime.ExtensionModule
@@ -1,7 +1,0 @@
-moduleName=OrbisGIS Groovy datastore extension
-moduleVersion=1.0
-extensionClasses=org.orbisgis.datastore.datastoreutils.DataStoreUtils, \
-  org.orbisgis.datastore.datastoreutils.FeatureIteratorUtils, \
-  org.orbisgis.datastore.datastoreutils.FeatureCollectionUtils, \
-  org.orbisgis.datastore.datastoreutils.FeatureSourceUtils
-staticExtensionClasses=org.orbisgis.datastore.datastoreutils.DataStoreFinderUtils
